Reorganizes internal package, removes services from common
