+++ internal/services/services_test.go
@@ -1,4 +1,4 @@
-package common
+package services
 
 import "testing"
