+++ rpc/modules/system.go
@@ -17,7 +17,7 @@
 package modules
 
 import (
-	api "github.com/ChainSafe/gossamer/internal"
+	"github.com/ChainSafe/gossamer/internal/api"
 	"net/http"
 )
